renaming file.ts to ownership.ts
