+++ src/lib/ownership/index.ts
@@ -1,5 +1,5 @@
 export { OwnedFile } from './lib/OwnedFile';
 export { OwnershipEngine } from './lib/OwnershipEngine';
-export { getOwnership } from './file';
+export { getOwnership } from './ownership';
 export { validate } from './validate';
 export { Matcher, FileOwnershipMatcher } from './types';
