+++ src/lib/ownership/ownership.ts
